requirements.txt
fastapi
uvicorn
fastai
pillow
